+++ mortgage_imports/sql/fannie/with_hpi_ct.sql
@@ -28,7 +28,7 @@ CREATE TABLE fannie.with_hpi
     borr_first_time_flg LowCardinality(FixedString(1)),
     ln_purp_cd LowCardinality(FixedString(1)),  
     prop_type_cd LowCardinality(FixedString(2)),
-    prop_num_unit Float32,
+    prop_num_unit Int16,
     prop_occ_cd LowCardinality(FixedString(1)),
     prop_st LowCardinality(FixedString(2)),
     prop_msa_cd LowCardinality(FixedString(5)),
